backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langgraph
langgraph-supervisor
neo4j 
pydantic
langsmith
python-dotenv
numpy
fastapi
numpy
openai
tenacity
markdown 
weasyprint 
python-docx
pandas
openai
sentence-transformers
py2neo
python-multipart